requirements.txt
Pillow
PyPDF2
cloudconvert
Flask==2.2.2
Flask-Cors==3.0.10
Werkzeug==2.2.3
groq
langchain
langchain_community
langchain_groq
numpy
opencv-python-headless
pdf2image
pdfplumber
playwright
plotly
pytesseract
python-docx
python-dotenv
pyvirtualdisplay
requests==2.28.2
scikit-learn
streamlit
streamlit_autorefresh
streamlit_lottie
openai
faiss-cpu
gunicorn==20.1.0
google-cloud-storage
docx2txt
beautifulsoup4
pandas
flake8
black
pre-commit
pytest-cov
redis[hiredis]
celery